Use state estimation on IEEE123 case
Now allows for different rates

See test_ieee123.json

+++ test_full_systems.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 # Load component framework
-from componentframework.basic_component import component_from_json
-from componentframework.system_configuration import (
+from gadal.componentframework.basic_component import component_from_json
+import argparse
+from os.path import isfile
+
+from gadal.componentframework.system_configuration import (
     generate_runner_config,
     WiringDiagram,
 )
@@ -26,6 +29,9 @@ StateEstimatorComponent = component_from_json(
 Recorder = component_from_json(
     "recorder/component_definition.json", bad_type_checker
 )
+IEEE123Feeder = component_from_json(
+    "dss_federate_123PV/component_definition.json", bad_type_checker
+)
 
 # Dictionary used to interpret test_system.json
 component_types = {
@@ -33,13 +39,21 @@ component_types = {
     "AWSFeeder": AWSFeeder,
     "MeasurementComponent": MeasurementComponent,
     "StateEstimatorComponent": StateEstimatorComponent,
-    "Recorder": Recorder
+    "Recorder": Recorder,
+    "IEEE123Feeder": IEEE123Feeder
 }
 
-# Read wiring diagram (lists components, links, and parameters)
-wiring_diagram = WiringDiagram.parse_file("test_system.json")
-#wiring_diagram.clean_model()
-# Generate runner config using wiring diagram and component types
-runner_config = generate_runner_config(wiring_diagram, component_types)
-with open("test_system_runner.json", "w") as f:
-    f.write(runner_config.json(indent=2))
+parser = argparse.ArgumentParser(description='Compiling wiring diagram json.')
+parser.add_argument('system_json', type=str,
+                    help='an integer for the accumulator')
+args = parser.parse_args()
+if isfile(args.system_json):
+    # Read wiring diagram (lists components, links, and parameters)
+    wiring_diagram = WiringDiagram.parse_file(args.system_json)
+    # Generate runner config using wiring diagram and component types
+    runner_config = generate_runner_config(wiring_diagram, component_types)
+    with open("test_system_runner.json", "w") as f:
+        f.write(runner_config.json(indent=2))
+        print("helics run --path=test_system_runner.json")
+else:
+    print(f"Wiring Diagram {args.system_json} not found")
